+++ src/env.rs
@@ -1,3 +1,4 @@
+use std::cell::RefCell;
 use std::collections::HashMap;
 use std::rc::Rc;
 
@@ -6,7 +7,7 @@ use crate::types::Type;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Env {
-    data: HashMap<String, Type>,
+    data: RefCell<HashMap<String, Type>>,
     outer: Option<Rc<Env>>,
 }
 
@@ -16,8 +17,8 @@ impl Env {
             panic!("`binds` and `exprs` must have the same length");
         }
 
-        let mut env = Env {
-            data: HashMap::new(),
+        let env = Env {
+            data: RefCell::new(HashMap::new()),
             outer: outer,
         };
         for i in 0..binds.len() {
@@ -27,7 +28,7 @@ impl Env {
     }
 
     pub fn new_default() -> Env {
-        let mut env = Env::new(None, &[], &[]);
+        let env = Env::new(None, &[], &[]);
         let ns = Namespace::new_default();
         for (sym, fun) in ns {
             env.set(&sym, Type::Fun(fun));
@@ -35,12 +36,12 @@ impl Env {
         env
     }
 
-    pub fn set(&mut self, symbol: &str, value: Type) {
-        self.data.insert(symbol.to_owned(), value);
+    pub fn set(&self, symbol: &str, value: Type) {
+        self.data.borrow_mut().insert(symbol.to_owned(), value);
     }
 
     fn find(&self, symbol: &str) -> Option<&Env> {
-        match self.data.get(symbol) {
+        match self.data.borrow().get(symbol) {
             Some(_) => Some(self),
             None => match self.outer {
                 Some(ref env) => env.find(symbol),
@@ -51,11 +52,31 @@ impl Env {
 
     pub fn get(&self, symbol: &str) -> Result<Type, String> {
         match self.find(symbol) {
-            Some(env) => match env.data.get(symbol) {
+            Some(env) => match env.data.borrow().get(symbol) {
                 Some(value) => Ok(value.clone()),
                 None => Err(format!("Env should have the symbol '{}'", symbol)),
             },
             None => Err(format!("Symbol '{}' not found in any environment", symbol)),
+        }
+    }
+
+    fn _outermost(self: &Rc<Env>) -> Option<Rc<Env>> {
+        match self.outer {
+            Some(ref outer) => {
+                let outermost = outer._outermost();
+                match outermost {
+                    Some(_) => outermost,
+                    None => Some(outer.clone()),
+                }
+            }
+            None => None
+        }
+    }
+
+    pub fn outermost(self: &Rc<Env>) -> Rc<Env> {
+        match self._outermost() {
+            Some(env) => env,
+            None => self.clone(),
         }
     }
 }